cleaning up components, need css on criteria screen
