--- app/src/components/topics/arithmetic/subtraction/SubtractionModule.jsx
@@ -1,9 +1,0 @@
-const SubtractionModule = (props) => {
-    return (
-        <>
-            <h1>Subtraction</h1>
-        </>
-    );
-}
-
-export default SubtractionModule;
